+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Current
 
-- TODO~~~~
+- Replace `ganache` with `anvil` for mainnet fork solution. Anvil is much more stable
+  than Ganache what comes to JSON-RPC. Anvil is much faster. You can now call
+  [fork_network_anvil]() that is a drop-in replacement for old
+  Ganache based `fork_network`.
 
 # 0.13.11